Updated Task System
Task Resources are no longer responsible for configuring a Task, tasks will now be provided a reference to the Task Resource they were spawned from, and are expected to be responsible for configuring themselves. Tasks will be provided a list of dependency Task nodes, based on a list of dependency Task IDs specified by the task.